mv isExpiring() tests to access-utils.spec
