+++ ui/src/app/utils/access-utils.spec.tsx
@@ -18,6 +18,7 @@ import {
   buildRasRedirectUrl,
   computeRenewalDisplayDates,
   getTwoFactorSetupUrl,
+  isExpiring,
   maybeDaysRemaining,
   NOTIFICATION_THRESHOLD_DAYS,
   RAS_CALLBACK_PATH,
@@ -421,5 +422,72 @@ describe('buildRasRedirectUrl', () => {
     expect(buildRasRedirectUrl()).toMatch(
       encodeURIComponent('http://localhost' + RAS_CALLBACK_PATH)
     );
+  });
+});
+
+describe('isExpiring', () => {
+  const ONE_MINUTE_IN_MILLIS = 1000 * 60;
+  const LOOKBACK_PERIOD = 99; // arbitrary for testing; actual prod value is 330
+  const EXPIRATION_DAYS = 123; // arbitrary for testing; actual prod value is 365
+
+  beforeEach(() => {
+    serverConfigStore.set({
+      config: {
+        ...defaultServerConfig,
+        accessRenewalLookback: LOOKBACK_PERIOD,
+      },
+    });
+  });
+
+  it('should return isExpiring=true if a module has expired', () => {
+    const completionDaysPast = 555; // arbitrary for test; completed this many days ago
+
+    // add 1 minute so we don't hit the boundary *exactly*
+    const completionDate =
+      nowPlusDays(-completionDaysPast) + ONE_MINUTE_IN_MILLIS;
+    const expirationDate = plusDays(completionDate, EXPIRATION_DAYS);
+
+    // sanity-check: this test is checking an expiration in the past
+    expect(expirationDate).toBeLessThan(Date.now());
+
+    expect(isExpiring(expirationDate)).toEqual(true);
+  });
+
+  it('should return isExpiring=true if a module will expire in the future and within the lookback period', () => {
+    const completionDaysPast = 44; // arbitrary for test; completed this many days ago
+
+    // add 1 minute so we don't hit the boundary *exactly*
+    const completionDate =
+      nowPlusDays(-completionDaysPast) + ONE_MINUTE_IN_MILLIS;
+    const expirationDate = plusDays(completionDate, EXPIRATION_DAYS);
+
+    // sanity-check: this test is checking a date within the lookback
+    const endOfLookback = nowPlusDays(LOOKBACK_PERIOD);
+    expect(expirationDate).toBeLessThan(endOfLookback);
+
+    expect(isExpiring(expirationDate)).toEqual(true);
+  });
+
+  it('should return isExpiring=false if a module will expire in the future beyond the lookback period', () => {
+    const completionDaysPast = 3; // arbitrary for test; completed this many days ago
+
+    // add 1 minute so we don't hit the boundary *exactly*
+    const completionDate =
+      nowPlusDays(-completionDaysPast) + ONE_MINUTE_IN_MILLIS;
+    const expirationDate = plusDays(completionDate, EXPIRATION_DAYS);
+
+    // sanity-check: this test is checking a date beyond the lookback
+    const endOfLookback = nowPlusDays(LOOKBACK_PERIOD);
+    expect(expirationDate).toBeGreaterThan(endOfLookback);
+
+    expect(isExpiring(expirationDate)).toEqual(false);
+  });
+
+  it('should return isExpiring=false if a module has a null expiration', () => {
+    expect(isExpiring(null)).toEqual(false);
+  });
+
+  it('should return isExpiring=false if a module has an undefined expiration', () => {
+    expect(isExpiring(undefined)).toEqual(false);
   });
 });
